feat(nodejs-base): export TracerBase without CLS and default plugin loads
