+++ packages/opencensus-nodejs-base/.npmignore
@@ -1,0 +1,4 @@
+/bin
+/coverage
+/doc
+/test
