+++ packages/opencensus-nodejs-base/AUTHORS
@@ -1,0 +1,8 @@
+# This is the list of OpenCensus authors for copyright purposes.
+#
+# This does not necessarily list everyone who has contributed code, since in
+# some cases, their employer may be the copyright holder.  To see the full list
+# of contributors, see the revision history in source control.
+
+Google LLC
+CESAR Team (www.cesar.org.br)
