+++ packages/opencensus-nodejs-base/codecov.yml
@@ -1,0 +1,2 @@
+fixes:
+  - "::packages/opencensus-nodejs/"
