+++ packages/opencensus-nodejs-base/scripts/copy-fixtures.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+find ./test -type d -exec mkdir -p ./build/{} \;
+find ./test -name 'package.json' -type f -exec cp {} ./build/{} \;
