+++ packages/opencensus-nodejs-base/src/trace/constants.ts
