+++ packages/opencensus-nodejs-base/src/trace/instrumentation/ext-types.d.ts
