+++ packages/opencensus-nodejs-base/tsconfig.json
@@ -1,0 +1,21 @@
+{
+  "extends": "./node_modules/gts/tsconfig-google.json",
+  "compilerOptions": {
+    "rootDir": ".",
+    "outDir": "build",
+    "pretty": true,
+    "module": "commonjs",
+    "target": "es6",
+    "strictNullChecks": true,
+    "noUnusedLocals": true,
+    "sourceMap": false
+  },
+  "include": [
+    "src/**/*.ts",
+    "test/**/*.ts",
+    "test/instrumentation/node_modules/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
